requirements.txt
# Frontend / App
streamlit
python-dotenv

# Pinecone (PINNED – VERY IMPORTANT)
pinecone[asyncio]>=6.0.0,<8.0.0
langchain-pinecone

# LangChain + LLM
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-google-genai

# Embeddings & Docs
sentence-transformers
pypdf
